cleaned some compiler warnings
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@3092 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/rsserver/rsinit.cc
@@ -210,7 +210,7 @@ void RsInit::InitRsConfig()
 
 #ifdef WINDOWS_SYS
 	// test for portable version
-	if (GetFileAttributes ("gpg.exe") != -1 && GetFileAttributes ("gpgme-w32spawn.exe") != -1) {
+        if (GetFileAttributes ("gpg.exe") != (DWORD) -1 && GetFileAttributes ("gpgme-w32spawn.exe") != (DWORD) -1) {
 		// use portable version
 		RsInitConfig::portable = true;
 	}
